CLOUDP-164756: Add missing tests for command builders
